+++ Source/DataSources/KmlDataSource.js
@@ -42,7 +42,6 @@ define([
         '../ThirdParty/zip',
         './BillboardGraphics',
         './CompositePositionProperty',
-        './CorridorGraphics',
         './DataSource',
         './DataSourceClock',
         './Entity',
@@ -108,7 +107,6 @@ define([
         zip,
         BillboardGraphics,
         CompositePositionProperty,
-        CorridorGraphics,
         DataSource,
         DataSourceClock,
         Entity,
@@ -1268,17 +1266,18 @@ define([
                 wall.outlineColor = defined(polygon.material) ? polygon.material.color : Color.WHITE;
             }
         } else if (dataSource._clampToGround && !canExtrude && tessellate) {
-            var corridor = new CorridorGraphics();
-            entity.corridor = corridor;
-            corridor.positions = coordinates;
+            var polylineGraphics = new PolylineGraphics();
+            polylineGraphics.clampToGround = true;
+            entity.polyline = polylineGraphics;
+            polylineGraphics.positions = coordinates;
             if (defined(polyline)) {
-                corridor.material = defined(polyline.material) ? polyline.material.color.getValue(Iso8601.MINIMUM_VALUE) : Color.WHITE;
-                corridor.width = defaultValue(polyline.width, 1.0);
+                polylineGraphics.material = defined(polyline.material) ? polyline.material.color.getValue(Iso8601.MINIMUM_VALUE) : Color.WHITE;
+                polylineGraphics.width = defaultValue(polyline.width, 1.0);
             } else {
-                corridor.material = Color.WHITE;
-                corridor.width = 1.0;
+                polylineGraphics.material = Color.WHITE;
+                polylineGraphics.width = 1.0;
             }
-            corridor.zIndex = zIndex;
+            polylineGraphics.zIndex = zIndex;
         } else {
             if (defined(zIndex)) {
                 oneTimeWarning('kml-gx:drawOrder', 'KML - gx:drawOrder is not supported in LineStrings when clampToGround is false');